+++ Plain Craft Launcher 2/Modules/Gamepad/ModGamepadInput.cs
@@ -5,7 +5,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using SharpDX.XInput;
 
-namespace PCL;
+namespace PCL.Modules.Gamepad;
 
 public static partial class ModGamepad
 {
